+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/__init__.py
@@ -22,6 +22,8 @@
 from ._container import ContainerProxy
 from ._cosmos_client import CosmosClient
 from ._database import DatabaseProxy
+from ._embedding_provider import EmbeddingProvider
+from .._embedding_result import EmbeddingResult
 from ._user import UserProxy
 from ._scripts import ScriptsProxy
 
@@ -30,5 +32,7 @@ __all__ = (
     "DatabaseProxy",
     "ContainerProxy",
     "ScriptsProxy",
-    "UserProxy"
+    "UserProxy",
+    "EmbeddingProvider",
+    "EmbeddingResult"
 )